Add a backends/ module, moving some files.
