+++ prototype/sky/backends/__init__.py
@@ -1,0 +1,7 @@
+from sky.backends.backend import Backend
+from sky.backends.cloud_vm_ray_backend import CloudVmRayBackend
+
+__all__ = [
+    'Backend',
+    'CloudVmRayBackend',
+]
